+++ includes/Gm2_Elementor_SEO.php
@@ -1,4 +1,7 @@
 <?php
+
+namespace Gm2;
+
 if (!defined('ABSPATH')) {
     exit;
 }
