+++ includes/Gm2_Tariff_Manager.php
@@ -1,4 +1,7 @@
 <?php
+
+namespace Gm2;
+
 if (!defined('ABSPATH')) {
     exit;
 }
